% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinfsc50-package.R
\docType{package}
\name{pinfsc50-package}
\alias{pinfsc50}
\alias{pinfsc50-package}
\title{pinfsc50: Sequence ('FASTA'), Annotation ('GFF') and Variants ('VCF') for 17 Samples of 'P. Infestans" and 1 'P. Mirabilis'}
\description{
Genomic data for the plant pathogen "Phytophthora infestans." It includes a variant file ('VCF'), a sequence file ('FASTA') and an annotation file ('GFF'). This package is intended to be used as example data for packages that work with genomic data.
}
\author{
\strong{Maintainer}: Brian J. Knaus \email{briank.lists@gmail.com}

Authors:
\itemize{
  \item Niklaus J. Grunwald \email{grunwaln@science.oregonstate.edu}
}

}
\keyword{internal}
